python/requirements.txt
# Python requirements for Resume Parser
# Install with: pip install -r requirements.txt

pdfplumber>=0.9.0
PyPDF2>=3.0.0
python-docx>=0.8.11
